levels
organized things into levels, I have 5 started.
I need to remove 4 cameras and change the script to teleport the camera when the player moves.
I need to design levels 3-5
I need to add the torch interaction
I need to add the mage prefabs (I'm thinking I'll have a slow moving prefab that soft follows the player, has like 0.1 mass w/rigid body so it kinda falls on an arc, and self destructs after something between 2-3 seconds).
Add the knight interaction.
The view is a bit small, so there's not much room to design the levels. So I might pull the camera back and shrink the objects, or do something to make the relative ratio of field to player larger so I can have more design freedom.